chore: Corrige tipo de id do usuario - Altera tipo int para long no Id do usuario - Melhora layout da tela de login

+++ XUnitBlog.Domain/Services/UserService.cs
@@ -13,7 +13,7 @@ public class UserService(
     IJwtService jwtService
 )
 {
-    public async Task<GetUserDto> GetById(int id)
+    public async Task<GetUserDto> GetById(long id)
     {
         var user = await repository.GetUserById(id);
